+++ src/main/java/nextstep/qna/domain/DeleteHistory/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.DeleteHistory;
 
 public enum ContentType {
     QUESTION, ANSWER;
