+++ src/main/java/nextstep/qna/domain/DeleteHistory/DeleteHistory.java
@@ -1,5 +1,7 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.DeleteHistory;
 
+import nextstep.qna.domain.answer.Answer;
+import nextstep.qna.domain.question.Question;
 import nextstep.users.domain.NsUser;
 
 import java.time.LocalDateTime;
