+++ src/main/java/nextstep/qna/domain/question/QuestionContents.java
@@ -1,5 +1,8 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.question;
 
+import nextstep.qna.domain.answer.Comments;
+import nextstep.qna.domain.answer.Answer;
+import nextstep.qna.domain.answer.Answers;
 import nextstep.users.domain.NsUser;
 
 import java.util.List;
