+++ src/test/java/nextstep/qna/domain/answer/AnswerTest.java
@@ -1,5 +1,6 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.answer;
 
+import nextstep.qna.domain.BaseEntity;
 import nextstep.users.domain.NsUser;
 import org.junit.jupiter.api.Test;
