+++ src/test/java/nextstep/qna/domain/deletehistory/DeleteHistoryTest.java
@@ -1,5 +1,13 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.deletehistory;
 
+import nextstep.qna.domain.BaseEntity;
+import nextstep.qna.domain.DeleteHistory.ContentType;
+import nextstep.qna.domain.DeleteHistory.DeleteHistory;
+import nextstep.qna.domain.answer.Answer;
+import nextstep.qna.domain.answer.Answers;
+import nextstep.qna.domain.answer.Comments;
+import nextstep.qna.domain.question.Question;
+import nextstep.qna.domain.question.QuestionContents;
 import nextstep.users.domain.NsUser;
 import org.junit.jupiter.api.Test;
