+++ src/test/java/nextstep/qna/domain/question/QuestionContentsTest.java
@@ -1,5 +1,7 @@
-package nextstep.qna.domain;
+package nextstep.qna.domain.question;
 
+import nextstep.qna.domain.answer.Answers;
+import nextstep.qna.domain.answer.Comments;
 import nextstep.users.domain.NsUser;
 import org.junit.jupiter.api.Test;
